ML Components/requirements.txt
tensorflow-cpu
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
streamlit
easyocr
scipy
scikit-learn
